+++ autofit/__init__.py
@@ -25,8 +25,9 @@ from .non_linear.samples import Sample
 from .non_linear.samples import load_from_table
 from .non_linear.samples import SamplesStored
 from .database.aggregator import Aggregator
-from .aggregator.aggregate_csv import AggregateCSV
-from .aggregator.aggregate_images import AggregateImages
+from .aggregator.summary.aggregate_csv import AggregateCSV
+from .aggregator.summary.aggregate_images import AggregateImages
+from .aggregator.summary.aggregate_fits import AggregateFits
 from .database.aggregator import Query
 from autofit.aggregator.fit_interface import Fit
 from .aggregator.search_output import SearchOutput
