+++ autofit/aggregator/summary/aggregate_fits.py
@@ -1,3 +1,3 @@
-class AggregateFits:
+class AggregateFITS:
     def __init__(self):
         pass
