+++ autofit/aggregator/summary/aggregate_images.py
